A vulnerable HSL Play in Scala app
It's a one-to-one translation of an (not yet?) open sourced HSL that
uses a Java API of Play framework.
So some constructs definitely still feel Java-like, even though we
already use Scala API.
It's a vulnerable app *on purpose* to provide a test bed for checking
static and runtime security analysis tools.
The app also uses a now deprecated API in Play, so we might want to
update that as well at some point.